pyproject.toml
[project]
name = "claude-blind-spots"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "tqdm>=4.67.1",
]
